+++ service/controller/rbac/resource/automation/delete.go
@@ -46,6 +46,7 @@ func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
 	clusterRoleBindings := []string{
 		pkgkey.WriteSilencesAutomationSAinNSRoleBindingName(ns.Name),
 		pkgkey.KamajiDatastoreManagerAutomationSAinNSRoleBindingName(ns.Name),
+		pkgkey.ClusterAppChartLookupsAutomationSAinNSRoleBindingName(ns.Name),
 	}
 
 	for _, clusterRoleBinding := range clusterRoleBindings {